Add Penn Treebank data processing pipeline
The goal is to have a reproducible way of going from official treebank
releases as distributed by LDC to the exact data files used to train the
parser. This data pipeline produces both the standard trees used in
publications, as well as trees with revised tokenization and syntactic
annotations that more closely match other treebanks and modern
tokenization conventions.

The data pipeline also recovers original whitespace (and non-escaped
text) from the raw text files underlying the treebank. This aims to
reduce the mismatch between parser training and pre-training (which uses
minimally processed unicode text), as well as downstream use in spaCy
(which also has non-destructive tokenization that preserves raw text).