Major re-write and cleanup of the pytorch code
Much of the custom code in this repository is better handled by a
library (some of these library features are new since the parser was
originally written and published):
- transformers now has fast tokenizers that return offset mappings
  between subword tokens and the input text. This allows replacing
  custom code for converting between PTB and BERT tokenization with a
  more general approach that uses raw text as a common format. A new
  data processing pipeline also allows using the raw WSJ text to
  determine where whitespace occurs between tokens, instead of having to
  guess and relying on heuristics that depend on implementation details
  of the BERT tokenizer.
- pytorch-struct provides a CKY primitive that allows implementing the
  margin loss used to train the parser. This means that the custom
  Cython implementation can be replaced with a library call, and the
  loss is now computed entirely on GPU without having to round-trip to
  the CPU and back.
- nltk provides a Tree object and functions for reading/writing trees to
  text files, so we can just use that instead of our own implementation.

This change also removes a few training options that are no longer
absolutely essential in light of recent advances in NLP. Pre-trained
transformer models tend to give better results than ELMo, and the
quality of models trained with randomly-initialized whole-word
embeddings and predicted part-of-speech tags doesn't justify the amount
of custom code these require. CharLSTM remains because it's still useful
in evaluation conditions where no pre-training is allowed.

This commit further includes general cleanup and re-formatting of the
code, as well as an update to run with pytorch 1.6 and python 3.8.

The new code breaks checkpoint compatibility with older releases, but
new models trained with this code achieve comparable results to
previous models.