Major rewrite of benepar package for v0.2.0a0
The package now directly uses the same pytorch modules and checkpoint
format as the training code. This also drops Python 2 support: the
minimum Python version required is now 3.7. The API is updated with new
functions, but it's largely backward compatible except for requiring a
new model format.

The original intent of exporting models to TensorFlow was to have a
stable on-disk format where the Python code would not need to be updated
for years at a time. This has not worked out in practice; in fact the
use of TensorFlow is barrier because I no longer use it in my own work
(and much of the research community has moved to pytorch, too). Unifying
around pytorch and the transformers library opens the way for publishing
better models that take advantage of advances in pre-training, as well
as allowing others to easily distribute compatible models for new
languages.